Classes-Ja
Classes in JavaScript provide a way to create reusable and organized code by encapsulating related data and behavior. They facilitate the creation of objects with consistent structure and functionality, promoting code reusability and modularity. Classes are an integral part of object-oriented programming in JavaScript, allowing developers to build complex applications with ease.